[OPENBUS-2241] Criação da ferramenta BusExplorer  - Correção de bug na atualização de tamanho de linha da tabelas com células multilinha  - Revisão do mecanismo de finalização da aplicação  - Revisão de realização de chamadas remotas dentro da EDT    -- instante quando se verifica se o usuário possui permissão de administração    -- construção das referências das facetas do barramento na BusAdminLib  - Pequena correção na BusAdmin  - atualização de entradas no LNG
git-svn-id: https://subversion.tecgraf.puc-rio.br/engdist/openbus/utils/busexplorer/trunk@148799 ae0415b3-e90b-0410-900d-d0be9363c56b

+++ BusAdminLib/src/main/java/admin/BusAdminImpl.java
@@ -49,7 +49,7 @@ public class BusAdminImpl implements BusAdmin {
   /** Host do barramento. */
   private String host;
   /** Porta do barramento. */
-  private short port;
+  private int port;
   /** {@link ORB} do barramento. */
   private ORB orb;
 
@@ -77,7 +77,7 @@ public class BusAdminImpl implements BusAdmin {
    * @param port Porta do barramento
    * @param orb ORB do barramento
    */
-  public BusAdminImpl(String host, short port, ORB orb) {
+  public BusAdminImpl(String host, int port, ORB orb) {
     connect(host, port, orb);
   }
 
@@ -89,7 +89,7 @@ public class BusAdminImpl implements BusAdmin {
    * @param port Porta do barramento
    * @param orb ORB do barramento
    */
-  public void connect(String host, short port, ORB orb) {
+  public void connect(String host, int port, ORB orb) {
     this.host = host;
     this.port = port;
     this.orb = orb;
